requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-crispy-forms==1.13.0
django-crum==0.7.9
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.1
pytz==2021.3
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
